Add links to visit future & past steps of a continuation (#77)
This PR provides links on the transaction detail pages (/txdetail) to step both forward and backward between individual steps of a continuation. At any given step, the user should be able to traverse back to any step before the current AND traverse to the immediate next step (even if it is fails). We also display failed next steps (we stop at 10). This PR completes the work done in both chainweb-data and chainweb-api for supporting this endeavor.

* Add links to visit future & past steps of a continuation

At this, we can visit previous steps from the current one

* Make definition table for past steps & initial code

* Links to step immediately after in continuation

* simplify display of failed conts

* Don't make an unnecessary query on the last step

* don't check if step is the last

* Update frontend/src/Frontend/Page/Common.hs

Co-authored-by: Enis Bayramoğlu <enisbayramoglu@gmail.com>

* ditch partial response from pactid endpoint

* Add line on page between failed and successful steps

* point to master on chainweb-api

* use span instead of partition b/c list is already sorted

---------

Co-authored-by: Enis Bayramoğlu <enisbayramoglu@gmail.com>