Add Signers and Signatures fields to txdetail view (#79)
This PR adds the "Signers" and "Signatures" fields to the `txdetail` view. These new fields display the new `TxDetails` fields of the `chainweb-data` API added by https://github.com/kadena-io/chainweb-data/pull/152 and https://github.com/kadena-io/chainweb-data/pull/153.

We used to have this information in the block transactions view before we redesigned that view to be a table of brief information about each transaction, each row linking to the corresponding `txdetail` page, losing the Signers and Signatures information in the process. This PR recovers these fields for the `block-explorer`.

* Add Signers and Signatures fields to txdetail view

* Remove Account and Scheme fields if they're empty

* Rename Account field to Address

* Separate the args of sig capabilities

* Let txDetailPage accept a NonEmpty list

* Rename txDetails

* Add separators to signatures