Flakeify block explorer (#80)
This PR adds a `flake.nix` layer around the legacy `default.nix` build setup for the `block-explorer`.

`block-explorer` relies on the excellent `obelisk` infrastructure for building the backend and the frontend. The version of `obelisk` depended on by `block-explorer` (and I believe latest version too) is not compatible with being imported into a flake out of the box. The reason is that `obelisk` (and its transitive dependencies) does impure things like looking up `builtins.currentSystem` and `import <nixpkgs>`, which won't work under flakes' pure evaluation. You can of course still include it in a flake and build the flake (and any downstream flake that depends on it) with `--impure`, but there are still 2 problems:
* `--impure` means that the resulting store hash depends on the environment in which the package is building, so you may get spurious cache misses
* The full Nix evaluation of `block-explorer` takes a non-trivial amount of time, and any downstream consumer of `block-explorer` will suffer from this additional evaluation time for every Nix action.
 
This PR addresses all of these issues by wrapping `default.nix` in a `recursive-nix` derivation, that makes the `nix-build -A exe` call as part of the build of its `defaultPackage`. This allows us to recover purity by sandboxing the environment in which Nix evaluation is taking place, so downstream consumers don't need `--impure`. Additionally, since all complex Nix evaluation is happening during the build of the `recursive-nix` derivation, evaluating the flake's `defaultPackage` is almost instantaneous (for any downstream consumer) and as long as the result is in the Nix store or is available in a binary cache, the whole build will be very fast. As long as there's a cache hit, downstream consumers won't even have to fetch any of the Nix dependencies of `block-explorer`.

One caveat with this wrapping is that, since it uses `recursive-nix`, any downstream consumer of this flake will need to enable `--extra-experimental-features recursive-nix` while building this flake, *but only if the result isn't found in a cache*. As long as the cache lookup succeeds, there's essentially no complication whatsoever.

One thing to note with the implementation of this PR is that I had to prepare [a long (and duplicate) list of nix tarballs](https://github.com/kadena-io/block-explorer/compare/enis/flakeify-block-explorer?expand=1#diff-370fdbf69abec4915ebf14c0bb9cfa0f38218732e223b018b2f244b49a2173fa) as [explicit dependencies of the `recursive-nix` derivation](https://github.com/kadena-io/block-explorer/pull/80/commits/830fd4ac6c2e8e0a538330b7863b47dddca2e5dc#diff-206b9ce276ab5971a2489d75eb1b12999d4bf3843b7988cbe8d687cfde61dea0R26). This was necessary, because `recursive-nix` derivations (like every sandboxed Nix derivation) don't have access to the internet, so they can't fetch those packages at Nix evaluation time (of the nested `nix-build` call). By making them explicit build dependencies of the `recursive-nix` derivation, we're making sure that at `nix-build` time they're already available. One interesting thing to note is that it's **OK** for the nested `nix-build` to need to fetch nix store paths from binary caches or access existing `/nix/store` paths on the local machine during its own **build time**, that's because of the way `recursive-nix` coordinates the actual build of any `.drv` generated during the recursive `nix-build`. The trouble is only with anything that needs to be fetched during the nested nix evaluation time.

This PR also post-processes the obelisk output to convert the result into a static SPA. This allows us to run the `block-explorer` on all platforms, even though the nix derivation only builds on `x86_64-linux`. Other platforms reuse the output produced by an `x86_64-linux`, which means they won't be able to build it themselves, but they will happily fetch it from a binary cache. The following shell command can be used to serve this SPA:
```
nix run github:kadena-io/block-explorer
```

* Add a flake wrapper around default.nix

* Extend the defaultPackage for other ${system}s

* Fix typo in systems list

* Remove aarch64-darwin

`block-explorer` doesn't seem to build on aarch64-darwin at the moment.

* Make block-explorer statically serveable

* Move evalTimeDependencies to separate file

* Make the server multi-platform

* Reduce copy-assets.sh noise

* Build and cache with Nix

* Make sure nix daemon has recursive-nix

* Restart daemon with sudo

* Reduce `static` closure

* Rename servePublic to serve

* Fix config encoding

* Fix root redirect

* Add default app

* Disable caching on index.html