Fix layout issues from wide cells (#90)
We're using nested tables for displaying complex nested data structures like blocks and transactions. The layout of this nested table structure gets distorted when a cell (potentially the cell of a nested table) becomes too wide and can't be line-wrapped.

This PR introduces a `tfieldLeaf` function as an alternative to `tfield`. `tfieldLeaf` allows us to make the "leaf" cells of these nested tables to be scrollable when needed.

Fixes #76 
Fixes #38